+++ src/submissions/frontend/all_window_functions.py
@@ -35,7 +35,10 @@ def select_save_file(obj:QMainWindow, default_name:str, extension:str) -> Path:
     Returns:
         Path: Path of file to be opened
     """    
-    home_dir = Path(obj.ctx["directory_path"]).joinpath(default_name).resolve().__str__()
+    try:
+        home_dir = Path(obj.ctx["directory_path"]).joinpath(default_name).resolve().__str__()
+    except FileNotFoundError:
+        home_dir = Path.home().resolve().__str__()
     fname = Path(QFileDialog.getSaveFileName(obj, "Save File", home_dir, filter = f"{extension}(*.{extension})")[0])
     return fname